refactor: split tools into individual modules, replace deprecated mcp.tool API
Replace monolithic src/tools.ts with src/tools/ directory containing one file
per tool. All tools now use mcp.registerTool() instead of deprecated mcp.tool().
Use proper moment default import with allowSyntheticDefaultImports.
