+++ src/tools/constants.ts
@@ -1,0 +1,3 @@
+export const ACCESS_DENIED_MSG = 'Access denied or resource not found.';
+export const MAX_SEARCH_RESULTS = 100;
+export const MAX_APPEND_LENGTH = 50000;
